Update backprop tests
